Moved and special config for Jest
