+++ frontend/src/components/calendar/__test__/calendar.test.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import Calendar from '../Calendar';
+
+it("Renders without problem", () => {
+    const div = document.createElement('div');
+    ReactDOM.render(<Calendar />, div);
+});
