+++ frontend/src/components/calendar/main.scss
